+++ samples/Sample_05_Chart.php
@@ -6,9 +6,9 @@ use PhpOffice\PhpPresentation\PhpPresentation;
 use PhpOffice\PhpPresentation\Shape\Chart\Type\Area;
 use PhpOffice\PhpPresentation\Shape\Chart\Type\Bar;
 use PhpOffice\PhpPresentation\Shape\Chart\Type\Bar3D;
-use PhpOffice\PhpPresentation\Shape\Chart\Type\Line;
 use PhpOffice\PhpPresentation\Shape\Chart\Type\Pie;
 use PhpOffice\PhpPresentation\Shape\Chart\Type\Pie3D;
+use PhpOffice\PhpPresentation\Shape\Chart\Type\Radar;
 use PhpOffice\PhpPresentation\Shape\Chart\Type\Scatter;
 use PhpOffice\PhpPresentation\Shape\Chart\Series;
 use PhpOffice\PhpPresentation\Style\Alignment;
@@ -537,7 +537,7 @@ function fnSlide_Pie(PhpPresentation $objPHPPresentation) {
     $shape->getLegend()->getFont()->setItalic(true);
 }
 
-function fnSlide_Scatter(PhpPresentation $objPHPPresentation) {
+function fnSlide_Radar(PhpPresentation $objPHPPresentation) {
     global $oFill;
     global $oShadow;
     
@@ -550,6 +550,62 @@ function fnSlide_Scatter(PhpPresentation $objPHPPresentation) {
     $seriesData = array('Monday' => 0.1, 'Tuesday' => 0.33333, 'Wednesday' => 0.4444, 'Thursday' => 0.5, 'Friday' => 0.4666, 'Saturday' => 0.3666, 'Sunday' => 0.1666);
     
     // Create a scatter chart (that should be inserted in a shape)
+    echo date('H:i:s') . ' Create a radar chart (that should be inserted in a chart shape)'.EOL;
+    $oChart = new Radar();
+    $series = new Series('Downloads', $seriesData);
+    $series->setShowSeriesName(true);
+    $series->getMarker()->setSymbol(\PhpOffice\PhpPresentation\Shape\Chart\Marker::SYMBOL_DASH);
+    $series->getMarker()->setSize(10);
+    $oChart->addSeries($series);
+    
+    // Create a shape (chart)
+    echo date('H:i:s') . ' Create a shape (chart)'.EOL;
+    $shape = $currentSlide->createChartShape();
+    $shape->setName('PHPPresentation Daily Download Distribution')
+        ->setResizeProportional(false)
+        ->setHeight(550)
+        ->setWidth(700)
+        ->setOffsetX(120)
+        ->setOffsetY(80);
+    $shape->setShadow($oShadow);
+    $shape->setFill($oFill);
+    $shape->getBorder()->setLineStyle(Border::LINE_SINGLE);
+    $shape->getTitle()->setText('PHPPresentation Daily Downloads');
+    $shape->getTitle()->getFont()->setItalic(true);
+    $shape->getPlotArea()->setType($oChart);
+    $shape->getView3D()->setRotationX(30);
+    $shape->getView3D()->setPerspective(30);
+    $shape->getLegend()->getBorder()->setLineStyle(Border::LINE_SINGLE);
+    $shape->getLegend()->getFont()->setItalic(true);
+
+    $oGridlines = new \PhpOffice\PhpPresentation\Shape\Chart\Gridlines();
+    $oGridlines->getOutline()
+        ->setWidth(1)
+        ->getFill()
+            ->setFillType(Fill::FILL_SOLID)
+            ->setStartColor(new Color(Color::COLOR_RED));
+    $shape->getPlotArea()->getAxisY()
+        ->setMajorGridlines($oGridlines)
+        ->getOutline()
+            ->setWidth(2)
+            ->getFill()
+                ->setFillType(Fill::FILL_SOLID)
+                ->setStartColor(new Color(Color::COLOR_BLUE));
+}
+
+function fnSlide_Scatter(PhpPresentation $objPHPPresentation) {
+    global $oFill;
+    global $oShadow;
+
+    // Create templated slide
+    echo EOL.date('H:i:s') . ' Create templated slide'.EOL;
+    $currentSlide = createTemplatedSlide($objPHPPresentation); // local function
+
+    // Generate sample data for fourth chart
+    echo date('H:i:s') . ' Generate sample data for chart'.EOL;
+    $seriesData = array('Monday' => 0.1, 'Tuesday' => 0.33333, 'Wednesday' => 0.4444, 'Thursday' => 0.5, 'Friday' => 0.4666, 'Saturday' => 0.3666, 'Sunday' => 0.1666);
+
+    // Create a scatter chart (that should be inserted in a shape)
     echo date('H:i:s') . ' Create a scatter chart (that should be inserted in a chart shape)'.EOL;
     $lineChart = new Scatter();
     $series = new Series('Downloads', $seriesData);
@@ -557,7 +613,7 @@ function fnSlide_Scatter(PhpPresentation $objPHPPresentation) {
     $series->getMarker()->setSymbol(\PhpOffice\PhpPresentation\Shape\Chart\Marker::SYMBOL_DASH);
     $series->getMarker()->setSize(10);
     $lineChart->addSeries($series);
-    
+
     // Create a shape (chart)
     echo date('H:i:s') . ' Create a shape (chart)'.EOL;
     $shape = $currentSlide->createChartShape();
@@ -623,6 +679,8 @@ fnSlide_Doughnut($objPHPPresentation);
 fnSlide_Pie3D($objPHPPresentation);
 
 fnSlide_Pie($objPHPPresentation);
+
+fnSlide_Radar($objPHPPresentation);
 
 fnSlide_Scatter($objPHPPresentation);